backend/requirements.txt
akshare==1.6.38
appdirs==1.4.4
asgiref==3.4.1
attrs==21.4.0
beautifulsoup4==4.11.1
cattrs==22.1.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.2
decorator==5.1.1
defusedxml==0.7.1
Django==3.2.9
django-cors-headers==3.11.0
django-cors-middleware==1.5.0
django-filter==21.1
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.7.2
djoser==2.1.0
drf-yasg==1.20.0
et-xmlfile==1.1.0
exceptiongroup==1.0.0rc8
greenlet==1.1.2
html5lib==1.1
httpie==2.6.0
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.1
jsonpath==0.82
lxml==4.9.1
MarkupSafe==2.1.1
numpy==1.23.0
oauthlib==3.2.0
openpyxl==3.0.10
packaging==21.3
pandas==1.4.3
py-mini-racer==0.6.0
pycparser==2.21
Pygments==2.10.0
PyJWT==2.4.0
PyMySQL==1.0.2
pyparsing==3.0.9
pypinyin==0.46.0
PySocks==1.7.1
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2021.3
requests==2.26.0
requests-cache==0.9.5
requests-oauthlib==1.3.1
requests-toolbelt==0.9.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.2.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.35
sqlparse==0.4.2
stripe==2.74.0
tabulate==0.8.10
tqdm==4.64.0
uritemplate==4.1.1
url-normalize==1.4.3
urllib3==1.26.7
webencodings==0.5.1
websocket-client==1.3.3
wincertstore==0.2
xlrd==2.0.1
